requirements.txt
cassandra-driver
fastapi
gunicorn
jupyter
pandas
pypyr
python-dotenv
tensorflow
uvicorn
boto3
dataclasses
pydantic-settings
pydantic
cryptography